Adds solver timeouts for feature model analyzer and slicer.

+++ plugins/de.ovgu.featureide.fm.core/src/de/ovgu/featureide/fm/core/analysis/cnf/solver/EmptySatSolver.java
@@ -242,9 +242,7 @@ public class EmptySatSolver implements ISatSolver {
 	public void reset() {}
 
 	@Override
-	public void setTimeout(int timeout) {
-
-	}
+	public void setTimeout(long timeout) {}
 
 	@Override
 	public int[] getInternalSolution() {
